Fixed title alignment
